--- packages/hds/src/components/TimeInput/index.tsx
@@ -1,1 +1,0 @@
-export * from './TimeInput';
