--- packages/hds/src/components/Toast/index.ts
@@ -1,1 +1,0 @@
-export * from './Toast';
