--- packages/hds/src/nextjs.ts
@@ -1,1 +1,0 @@
-export * from '@chakra-ui/next-js';
